Tests created.Fixing problem with different result of parallel and linear algorithms

+++ src/main/java/TTL/controllers/datainspector/encoder/LabelEncoder.java
@@ -5,5 +5,5 @@ import TTL.models.Order;
 import java.util.List;
 
 public interface LabelEncoder {
-    int[] labelEncode(List<Order> orders);
+    double[] labelEncode(List<Order> orders);
 }
